refactor(chantiers): extraire resolveResponsables dans un fichier partagé
